extract the test results upload
